@@ -1,5 +1,5 @@
 {
-  "name": "service-name",
+  "name": "mapproxinator",
   "version": "1.0.5",
   "description": "This is template for map colonies typescript service",
   "main": "./src/index.ts",
@@ -38,59 +38,59 @@
     }
   },
   "dependencies": {
-    "@godaddy/terminus": "4.9.0",
-    "@map-colonies/error-express-handler": "^2.0.0",
-    "@map-colonies/express-access-log-middleware": "^0.0.3",
-    "@map-colonies/js-logger": "^0.0.3",
-    "@map-colonies/openapi-express-viewer": "^2.0.1",
-    "@map-colonies/telemetry": "3.0.0",
-    "@opentelemetry/api": "1.0.1",
+    "@godaddy/terminus": "^4.11.2",
+    "@map-colonies/error-express-handler": "^2.1.0",
+    "@map-colonies/express-access-log-middleware": "^2.0.1",
+    "@map-colonies/js-logger": "^1.0.1",
+    "@map-colonies/openapi-express-viewer": "^3.0.0",
+    "@map-colonies/telemetry": "4.1.0",
+    "@opentelemetry/api": "1.4.0",
     "@opentelemetry/api-metrics": "0.23.0",
-    "@opentelemetry/instrumentation-express": "0.23.0",
-    "@opentelemetry/instrumentation-http": "0.23.0",
+    "@opentelemetry/instrumentation-express": "0.32.1",
+    "@opentelemetry/instrumentation-http": "0.35.1",
     "aws-sdk": "^2.951.0",
     "compression": "^1.7.4",
-    "config": "^3.3.6",
-    "express": "^4.17.1",
-    "express-openapi-validator": "^4.12.14",
-    "http-status-codes": "^2.1.4",
+    "config": "^3.3.9",
+    "express": "^4.18.2",
+    "express-openapi-validator": "^5.0.3",
+    "http-status-codes": "^2.2.0",
     "pg": "^8.6.0",
     "reflect-metadata": "^0.1.13",
     "swagger-ui-express": "^4.1.6",
-    "tsyringe": "^4.5.0"
+    "tsyringe": "^4.7.0"
   },
   "devDependencies": {
-    "@commitlint/cli": "^9.1.1",
-    "@commitlint/config-conventional": "^10.0.0",
-    "@map-colonies/eslint-config": "^2.1.0",
+    "@commitlint/cli": "^17.4.4",
+    "@commitlint/config-conventional": "^17.4.4",
+    "@map-colonies/eslint-config": "^3.1.3",
     "@map-colonies/prettier-config": "0.0.1",
     "@map-colonies/standard-version-update-helm-version": "^2.0.1",
-    "@map-colonies/standard-version-update-openapi3-version": "0.0.1",
-    "@redocly/openapi-cli": "^1.0.0-beta.25",
-    "@types/compression": "^1.7.0",
-    "@types/config": "0.0.36",
-    "@types/express": "^4.17.12",
-    "@types/jest": "^26.0.19",
+    "@map-colonies/standard-version-update-openapi3-version": "0.0.2",
+    "@redocly/openapi-cli": "^1.0.0-beta.94",
+    "@types/compression": "^1.7.2",
+    "@types/config": "^3.3.0",
+    "@types/express": "^4.17.17",
+    "@types/jest": "^29.4.0",
     "@types/js-yaml": "^4.0.1",
-    "@types/multer": "^1.4.5",
+    "@types/multer": "^1.4.7",
     "@types/pg": "^8.6.1",
-    "@types/supertest": "^2.0.10",
-    "@types/swagger-ui-express": "^4.1.2",
-    "commitlint": "^9.1.2",
-    "copyfiles": "^2.3.0",
-    "cz-conventional-changelog": "^3.0.1",
-    "eslint": "^7.8.1",
-    "husky": "^4.2.5",
-    "jest": "^26.6.3",
-    "jest-create-mock-instance": "^1.1.0",
-    "jest-html-reporters": "^2.0.3",
+    "@types/supertest": "^2.0.12",
+    "@types/swagger-ui-express": "^4.1.3",
+    "commitlint": "^17.4.4",
+    "copyfiles": "^2.4.1",
+    "cz-conventional-changelog": "^3.3.0",
+    "eslint": "^8.36.0",
+    "husky": "^8.0.3",
+    "jest": "^29.5.0",
+    "jest-create-mock-instance": "^2.0.0",
+    "jest-html-reporters": "^3.1.4",
     "js-yaml": "^4.1.0",
-    "prettier": "^2.2.1",
-    "pretty-quick": "^3.1.0",
-    "rimraf": "^3.0.2",
-    "standard-version": "^9.3.0",
-    "supertest": "^6.0.1",
-    "ts-jest": "^26.3.0",
-    "typescript": "^4.2.4"
+    "prettier": "^2.8.4",
+    "pretty-quick": "^3.1.3",
+    "rimraf": "^4.4.0",
+    "standard-version": "^9.5.0",
+    "supertest": "^6.3.3",
+    "ts-jest": "^29.0.5",
+    "typescript": "^4.9.5"
   }
 }
