@@ -25,7 +25,7 @@ export function convertJsonToYaml(json: Record<string, unknown>): string {
   try {
     const yamlContent: string = dump(json, { noArrayIndent: true });
     return yamlContent;
-  } catch (error) {
+  } catch (error: any) {
     throw new Error(error);
   }
 }
