@@ -5,7 +5,7 @@ import { createServer } from 'http';
 import { createTerminus } from '@godaddy/terminus';
 import { Logger } from '@map-colonies/js-logger';
 import { container } from 'tsyringe';
-import { get } from 'config';
+// import { get } from 'config';
 import { DEFAULT_SERVER_PORT, Services } from './common/constants';
 import { getApp } from './app';
 import { IConfig } from './common/interfaces';
@@ -27,7 +27,7 @@ const pollManager = container.resolve(PollManager);
 const initMode = config.get<boolean>('initMode');
 
 const startServer = (): void => {
-  const serverConfig = get<IServerConfig>('server');
+  const serverConfig = config.get<IServerConfig>('server');
   const port: number = parseInt(serverConfig.port) || DEFAULT_SERVER_PORT;
   const app = getApp();
   const healthCheck = container.resolve(Liveness).probe;
